Allow to specify parameter/return/update path in .pysa models
Summary:
This diff implements a new syntax in taint stubs that allow specifying the parameter, return and update path for sources, sinks and taint-in-taint-out.
The path must be specified using the access path syntax: It must start with `_` and be followed by: attribute accesses `_.foo`, index accesses `_["foo"]` or `_[0]`, or calls: `_.all()` which represents all taint on any attribute and `_.keys()` which represents all taint on dictionary keys.
Paths can be specified with `ParameterPath[...]`, `ReturnPath[...]` and/or `UpdatePath[...]` (for `Updates`).

We will add documentation in a follow-up.

Reviewed By: alexkassil

Differential Revision: D40705438

fbshipit-source-id: 9c15a8df27b14539632c051a363fa646ef200c48