remove name field from subscription responses on the client
Summary:
Context: The classic persistent subscription protocol includes a subscription name that's ferried back-and-forth from the server, and the client only handles requests that match the subscription name. However, the subscription connection actually already serves as a unique identifier for a connection, so we don't need to keep the name around.

With the upcoming codenav work, the subscriptions don't have names, so simplify the response structure on the client side to be able to use the unified structure between the two in the subscription loop.

Reviewed By: grievejia

Differential Revision: D40746468

fbshipit-source-id: 4b9eb4071e3159f68e6859cbeb0a340bac1f87db