start parsing subscription messages from codenav server
Summary:
This diff pulls out the component that parses subscription messages returned from a server and adds support for the code navigation server's flavor which doesn't include a subscription name. Shares as much code as possible with the existing subscription parsing.

Note that I chose to create a separate component with a single abstract method responsible for parsing for ease of understanding (could also send an arbitrary callable, but this made more sense to me...)

Reviewed By: joshuaxu1995

Differential Revision: D40746036

fbshipit-source-id: 580670365676f786c33d51ab80342213269de486