Use pyre flavors to determine the log directory for servers
Summary:
Context:

We were hard-coding the "new_server" name in server creation/rages/etc. With the imminent codenav server, the server's logs should be in a different directory, so let's transfer control of the logs to the flavor itself.

One suggested change was to have the flavor name be identical to the log directory name. I'm not opposed to this, but was concerned about two things:

1. backwards compatibility - things might depend on Pyre servers being in a `new_server` subdirectory, and didn't want to break the world due to an abstract fear of change.

2. For things like the classic codenav flavor and shadow servers, the same log directory as old servers might make sense. Do we want these to be in different dirs?

Reviewed By: grievejia

Differential Revision: D40729862

fbshipit-source-id: 7ffa7c9e03fd19244da2429c64dd87e531c883d9