create a base class for language server daemons
Summary:
Context: We will initially be able to share most of the policy between daemons of the persistent client and the codenav client. Create an abstract class that the persistent daemon extends from.

This is the last bit before we can start hooking everything up and have the codenav client spawn a server + do the initialization handshake.

Reviewed By: grievejia

Differential Revision: D40692754

fbshipit-source-id: c9fbb6c81a4712f645a2947f360b2462a74d736b