Use ParameterPath/ReturnPath instead of AppliesTo in integration tests
Summary:
Now that we introduced `ParameterPath` and `ReturnPath`, we can deprecate the use of `AppliesTo` since it does the same thing but is less flexible.
This diff removes `AppliesTo` from most of our tests (except `applies_to.py`).
This also removes a broken test that combines `AppliesTo` with `AttachXXX`, which was never properly supported.

Reviewed By: alexkassil

Differential Revision: D40744424

fbshipit-source-id: 920a816517c95b51817122803aa377acb97740b6