Fix a performance regression during control flow graph construction
Summary: Our control flow graph construction is quadratic in the number of statements, which can be very slow for auto-generated functions with hundred of thousands of lines of code. This is due to inserting at the end of a linked list, for every new statement. This can be easily fixed by inserting in the front and performing a list reversal at the end.

Reviewed By: onionymous

Differential Revision: D40832175

fbshipit-source-id: 7a614095ffdfb0331eeb10102f617fbfb32507e7