Add clause to set the state of opened documents even if overlays are off
Summary:
I'm seeing an issue right now where the `is_dirty` flag from the Pyre side and Pyls side is completely different. Additionally, it seemed like the flag set by Pyre was consistently less often to be true than for Pyls. I was puzzled why that occurred until I dived a little deeper into the code.

What I saw here was that for all non-overlay enabled users, the dirty flag would be consistently off for Pyre, and then I discovered my bug which was that in the case that `process_unsaved_changes` is False, we're not setting the `is_dirty` flag for `didChange` events. We need to do that here.

Reviewed By: stroxler

Differential Revision: D40915313

fbshipit-source-id: 3ec019278dd92309cb47195578f9a97d17c9a303