Document EnvironmentTable.In
Summary:
Understanding the internals of EnvironmentTable isn't
really needed to be productive working with environments,
unless you are actually trying to change the physical
representation.

The most important thing is actually to just understand
how the `In` implementation will map to behavior of that
layer.

As a result, I decided to start working on `environment.ml`
docs by focusing on documenting the `In` signature:
- what are `Key` and `Value`, and how is missing data
  typically represented?
- what is this mysterious `trigger` nonsense?
- why do we need `trigger <-> dependency` mappings, and
  why is that map injective rather than bijective?
- what does overlay ownership do?

Reviewed By: joshuaxu1995

Differential Revision: D40896178

fbshipit-source-id: 8465dd3f894808418def44dd34246adbace7554a