Fix some weird formatting in hack_parallel function signatures
Summary:
I was skimming this code to understand the limits of our
shared-memory multiprocessing-based concurrency logic, and
the fact that we had two args on one line was pretty distracting
(note that we don't autoformat this directory because it is
legacy code and the comments can't really handle it).

Just so I can skim the code, let's put each argument on one line,
and the return type if any on a separate line as well.

Reviewed By: grievejia

Differential Revision: D40960996

fbshipit-source-id: 78967d20e056cd97a5ba5ca3e3881c84a9178384