parse hover requests in the client and thread to server
Summary:
This diff starts parsing hover requests and creates one-off connections to a given CodeNav server to exercise the hover capability.

The main ideas are:

1. We have a code_navigation_request module that holds the requests intended to be passed into a code nav server.

2. We have an API that handles serializing a codenav request and passing that to a server, and reading results back (not yet handled: parsing the response).

As a consequence of the other changes, I had to move `daemon_connection` to be in the language_server directory for import purposes. Seems like it fits better there, anyway :)

The parsing of responses is out of scope for this diff, coming up next.

Reviewed By: pradeep90

Differential Revision: D40908604

fbshipit-source-id: 716d21c1a1476a794e3dca415e0785370b426560