change format of LocationOfDefinition response to be a record
Summary:
Context: On the client side, we want to parse LocationOfDefinition responses as a dataclass. However, the serialization of LocationOfDefinition as is
takes the shape of `["LocationOfDefinition", [d1, d2]]`, which can't be parsed as a dataclass. Instead having it be `{"definitions": [d1, d2]}` makes it natural to have a
uniform definition that's easy to read.

Reviewed By: grievejia

Differential Revision: D40955213

fbshipit-source-id: c2e0068897253bf78782f14de44c58ad35459e14