Print error if parse error is returned
Summary:
**TDLR**: Return presence of parse errors where a specific module has a parse error, to better understand cases in which an error recoverable parser would improve the accuracy of our definition requests.

We currently are doing an analysis on whether unsaved changes would improve the accuracy of our definition requests. In addition to unsaved changes, the error recoverable parser may also help. We want to collect early data on whether a specific module has a parse error because we can use this data for telemetry purposes, to forecast whether the unsaved changes + ERP is going to help our definition accuracy.

Design Choices:

A. **Server or Clientside implementation**

Since the data of parsing error is only for telemetry purposes, it would potentially be ideal that we could do this quickly and hackily on the clientside. However, the only I could think of doing this on the clientside is with the Python ast library (via `ast.parse`) Unfortunately, that has the potential of blocking our client, because the `ast.parse` is a synchronous method with synchronous internals that we cant easily asyncify.

Because there wasn't a quick way to do this on the client,  I decided to do it on the server.

B. **Where to make this change?**

I wanted to make the change in a high level module like `query.ml` to reduce cascading effects of changing many tests.

Reviewed By: stroxler

Differential Revision: D40984414

fbshipit-source-id: 912306cd193598c08c47aadd0c208353112496a1