Change Overlay Requests to use the new implementation of overlay id (which includes process id)
Summary:
**TLDR**: There's a bug right now in which definition requests are not getting responses back even for overlays. This diff fixes that bug.

The current issue is that I previously made a change that changes the overlay_id of overlay updates, but didn't make the corresponding change for definition requests. This causes all definition requests with overlays to fail, which explains why Pyre accuracy is so poor for dirty files, when overlays are enabled. This change will likely really improve our pyre accuracy #s.

Reviewed By: stroxler

Differential Revision: D40997758

fbshipit-source-id: 44f99dd6a112745cd482a407dbb0df5b620974c4