Use the call graph to check if a call is a mapping method
Summary:
Instead of using type resolution during the taint analysis (which can be costly), let's use the call graph information (which is already based on type resolution) to check whether a method call is a mapping or dictionary method.

This also has the benefit of fixing a false positive with `TypedDict`.

Reviewed By: alexkassil

Differential Revision: D40917926

fbshipit-source-id: 2b3746d8d447efd195adc0223d8b8d397bc89bd4