Fix a false negative wih dict.update and keyword arguments
Summary:
We are currently handling `d.update({"a": 1})` and `d.update(x={"a": 1})` similarly, even though the second call creates a key `x` instead of `a`.
This diff fixes it by matching on positional arguments only. I used this opportunity to do the same for most of our hardcoded behaviors, to prevent similar bugs.

Reviewed By: alexkassil

Differential Revision: D40950056

fbshipit-source-id: f3e1842d1a41e1db54639bf4a5b7def5b37e11c6