Hardcode the behavior of __setitem__ for sequences
Summary:
We already hardcode the behavior of `__setitem__` for dictionaries. It doesn't cost us much to also handle it for sequences (including `list`). This should improve the precision in some cases.

This also fixes a discrepancy between our internal tests and the actual analysis. Before D40921141, `list.__setitem__` calls would not be resolved and thus would be using the hardcoded (precise) behavior for `dict.__setitem__`. The actual analysis would have false positives since `list.__setitem__` would be properly resolved. See the example in the integration test.

Reviewed By: alexkassil

Differential Revision: D40921458

fbshipit-source-id: 9102f1e3e9b9e9209c1749adf60a70df5112e1ec