Treat object.__init__ as a no-op
Summary:
This is a pretty subtle change. I will try to explain it as best as I can.

# Context
We currently treat `object.__init__` as an obscure call, which means it propagates taint from its argument to `self`. This is technically wrong since `object.__init__` does not accept any parameters, hence it is a no-op from a taint perspective, and this diff aims to fix that.

The reason we currently treat `object.__init__` as an obscure call is to properly model class stubs, such as `class T: ...`, which means `T(x)` will be resolved as calling `object.__new__` and `object.__init__`. In that specific case, we want to taint the result with the taint of `x`, since we are missing the code of the class and anything could happen.

# Problem
The current behavior is a blocker to better model container constructors.
For instance, `frozenset` has a custom `__new__` but uses the default `object.__init__`. If we model `__new__` properly, it wouldn't matter since we would call `object.__init__` next, which will be an obscure call and lead to taint-in-taint-out with collapsing.

# Solution
In this diff, we treat `object.__init__` as a no-op, as it should be.
To properly handle class stubs, we special case expressions that call both `object.__new__` and `object.__init__`.

Reviewed By: alexkassil

Differential Revision: D41010190

fbshipit-source-id: 20752b4d976fc72628269d39b862885ffa833254