Fix models for standard functions taking dictionaries as iterables
Summary:
In Python, dictionaries are iterables and return the set of keys. For instance, `list({'a': 0, 'b': 1}) == ['a', 'b']`.

In our models for standard containers, we often use `ParameterPath[_.all()], ReturnPath[_.all()]`, which is a way to collapse the taint at one level. This performs a read of `AnyIndex` and prepends the result with `AnyIndex`. This basically results in a "blur" of one level in the taint tree, for instance:
```
# l -> {0: UserControlled, 1: Header}
l.sort()
# l -> {*: {UserControlled, Header}}
```

The problem is that `.all()` (i.e, `AnyIndex`) does not include `.keys()` (i.e, `AccessPath.dictionary_keys`). Hence, using `ParameterPath[_.all()], ReturnPath[_.all()]` will not propagate dictionary keys.
In this diff, we add a `TaintInTaintOut` for `ParameterPath[_.keys()]` on all container methods taking arbitrary iterators, to fix the false negative.

Reviewed By: alexkassil

Differential Revision: D41006538

fbshipit-source-id: 8d0c687610049c771e2df41b8a420e6ebf6257b5