Rename slave -> ephemeral worker
Summary:
An emphemeral worker is a daemon that is forked
off of a long-lived worker that lives for the duration
of a single request of multiprocessing work.

The double-fork approach here (spawn a worker, then
fork an ephemeral worker from it) is clearer if we use
the name "ephemeral worker".

My main motivation though is just that we ideally don't
want to be using slave as a name for any abstraction
in our codebase; I didn't realize this existed until
a few days ago.

Reviewed By: alexkassil

Differential Revision: D41000034

fbshipit-source-id: cbb8629306a3f560404866ebd9d6d8297d00de89