add ability to stop codenav servers
Summary:
Add a `--flavor` argument to the command.

Only design choice here is to associate the `--flavor` with the stop command instead of putting it on `pyre --flavor`. The idea is that only the server commands (start, stop, restart, servers) will need this, so better to constrain it to the places where we know it's useful.

For now, I'm leaving restart alone (we don't yet provide a way to start a codenav server without going through the `pyre code-navigation` command, and I'm not sure we have a great reason to support one?)

Reviewed By: pradeep90

Differential Revision: D41043884

fbshipit-source-id: ccde6713be5c58cf01a74c78388f6696d047572a