Fix Issue in DidChange Event that causes Definition Requests to Fail to get Correct Overlay Id
Summary:
**TLDR**: Fixing a bug that was caused by resolve() method.

In didChange events, I was using the resolve() method on document_path, which can cause the document_path to be expressed using a different path: if you see in this paste: P545919826, the resolve method resolves the /home/jxu425 directory to /data/users/jxu425. As a result, it causes inconsistent paths between the overlay and the definition request, which caused a definition request to not use the correct overlay environment.

Reviewed By: sinancepel

Differential Revision: D41043234

fbshipit-source-id: 7c87b1b413158ac466866fe4f46217b0698a6a65