Demonstrate a bug in call graph construction for TypedDict __init__
Summary:
The current call graph construction returns `.__init__` (with an empty class name) when creating an instance of a `TypedDict`. This diff demonstrates the bug.
Besides the call graph being wrong, this is still sound since it is treated as an obscure call. It just prevents us from providing better models for `__init__` of `TypedDict`s.
This will be fixed in a follow-up.

Reviewed By: alexkassil

Differential Revision: D41055276

fbshipit-source-id: 1c9673ad3a65122d58573d70fb0a544e56cce968