Support f-strings.
Summary:
Problem: We need to check the various substrings of the f-string.

Note that the result is always `Mutable`, just like strings, o that it can be assigned to `Mutable` or `ReadOnly`. Otherwise, if we naively mark it as `ReadOnly`, then we wouldn't be able to do `s: str = f"..."`.

RFC in case people think `Mutable` is more descriptive than `bottom` here.

Reviewed By: sinancepel

Differential Revision: D40878067

fbshipit-source-id: ef961a52b18e950a4f5de01fc0e93dc8e2a7cbb7