Bulletproof the handling of unsaved changes availablilty
Summary:
Prior to this commit, we were attempting to use implicit global
state in the OpendDocumentState struct to handle whether
unsaved-changes are enabled or disabled.

This used to work okay when we just had a didChange handler, that
was only turned on when unsaved changes were enabled. But today
we also use didChange to help with uptime tracking, and as a
result trying to interpret the unsaved-changes availability
there is quite error prone. I think we actually had a slight
bug where didSave and didOpen requests could potentially schedule
an overlay update even when unsaved changes should be off!

In this diff, I put the handling in what seems like the logical
place that makes reasoning about flags simpler:
- didChange / didOpen / didSave handlers track state for
  telemetry and do not check for unsaved-changes at all
- the actual overlay function, the method
  `update_overlay_with_latest_code`, checks for unsaved
  changes support

This makes it much easier to see that
 (a) telemetry should be lined up when the flag is off/on
 (b) we'll never accidentally update overlays when the flag
     is supposed to be off

Reviewed By: grievejia

Differential Revision: D41130448

fbshipit-source-id: 35bdc07c51c02ed9c32345e0f9ee984c29c86ae7