Move update-overlay method
Summary:
In principle, the update-overlay functionality is
generic
- All of the handlers that hit the daemon may eventually
  need to use it, including the didChange handler if we
  want to do type checking for unsaved changes.
- in particular, I want to use it on hover as well as
  definition requests (hover is my ideal case for manual
  testing because it's the easiest to check, just mouse over
  something after inserting a newline!).

So, let's move it up to the top of the server class
so that various handlers can easily use it.

An added benefit is that the helper method for definition
requests is once again directly above the main method; the
fact that we had an overlay update handler in between was
making it harder to see the logic, which matters because
I would like to refactor the `process_definition_request`
method.

Reviewed By: grievejia

Differential Revision: D41131208

fbshipit-source-id: adca093d9295011efd9aea7fd7afefa07727c44a