Rename update-overlay method
Summary:
I want it to be obvious at callsites that this method
may or may not actually update an overlay, because otherwise
someone reading the code is going to be worried about whether
they need an if/else block guarding the call.

Since it's pretty obvious that we're updating with the latest
code (what else would we upate with?) I decided to shorten
the name rather than having a very long name.

Reviewed By: grievejia

Differential Revision: D41138152

fbshipit-source-id: b9170ecd259cdd7a285424cba05946c853e9d578