Unsaved Changes outside Shadow: hover requests
Summary:
This commit enables non-shadow-mode dogfooding of Pyre
hover results with unsaved changes support.

Getting this in place has been a big goal for me because
it is so easy to manually end-to-end test hover, and I
can therefore get super clear signal as to whether
unsaved changes is working properly.

Reviewed By: grievejia

Differential Revision: D41138109

fbshipit-source-id: 9622fe71778620f7eeb38d267a11197e96b9b681