Refactor: deduplicate core definition logic
Summary:
This commit is at the heart of the cleanups I wanted to
do in the `process_definition_request` of `PyreLanguageServer`.

It does not solve all of the problems, but it solves what
I consider the most urgent problem: making sure that the
code I can easily end-to-end test by dogfooding Pyre (the
non-shadow-mode case) is running the same overlay update
and query logic that we're going to gather analytics
data on (the shadow mode case).

The other problems are just cosmetics and readability, which
is important to me but represents far less risk in my mind.

Reviewed By: sinancepel

Differential Revision: D41170266

fbshipit-source-id: befbad5cfb494a12d9b15f2a3f37c93c3ec7d076