Simplify definition handler: early return + dedent
Summary:
In the case where a file isn't part of the opened document set,
there's nothing for Pyre to do and no real need to make sure
the daemon is up (which was the only thing in the remainder
of `process_definition_request` that ran in that case).

So let's simplify with an early return and a dedent, it's an
easy readability improvement.

Reviewed By: sinancepel

Differential Revision: D41190375

fbshipit-source-id: 235bf12e49a9ff72c097c0659104a4dd1695567f