resolve paths in codenav requests
Summary: We use pathlib's `.resolve()` functionality to spawn codenav servers on the directory where a `.pyre_configuration.codenav` exists. This has the side effect of resolving all symlinks, i.e. if some has, say, `/a` as a symlink to `/b/c`, a server that's spawned under `/a/username` will actually spawn under `/b/c/username`. However, VSCode will still send paths for `/a/username`, so we need to translate these somehow. For now, I chose to resolve all symlinks fully as that's a consistent solution, but might incur a bunch of overhead over the long run.

Reviewed By: stroxler

Differential Revision: D41198818

fbshipit-source-id: dc94b58dcc5fe3ed21ed8f06e23920ea658d7201