in LocationBasedLookupProcessor, s/environment/type_environment
Summary:
I was looking into how to best support unsaved changes for type errors,
and one possible route was to put it into Query.ml. An issue I ran into
is that the TypeEnvironment.t value is named environment, which isn't
very explicit and would get confusing if I wound up passing in an
ErrorsEnvironment.t.

I think I may wind up going a different route for the unsaved changes
implementation, but since I was breifly blocked I wanted to go ahead
and rename things so that it's more obvious *which* environment we're
dealing with here.

Splitting it into 2 diffs to keep the line counts reasonable.

Reviewed By: grievejia

Differential Revision: D41207355

fbshipit-source-id: 77d473c698140cda1f5cd2a9c166b4f8a65aa6b4