Remove misleading log line
Summary:
We don't really need to log that we aren't updating the overlay,
that's possible to see from the absence of request / response logs.

More importantly, the log line was actually misleading because most
users don't have unsaved changes enabled in classic-mode servers
which made it look like there was a state handling bug in configerator.

Reviewed By: kinto0

Differential Revision: D41271664

fbshipit-source-id: 9f3864140c588f814091ecb1c44fdb3a99089e33