Add overlay type errors request to the backend
Summary:
All we need is the ability to grab type errors for one path
in an overlay. I talked this through with Jia and we think that
avoiding mutating the existing request variants is best; changing
them is quite challenging because it can break backward compatibility
of the serialization.

Since there's no backward compatibility issue, adding this as a new request type
is zero-risk.

Reviewed By: sinancepel

Differential Revision: D41269880

fbshipit-source-id: 2847ca647457a28afacb73a22b7016ea143eeafc