TreeDomain: added transform_on_sink & transform_on_hoist functions
Summary:
TreeDomain has a built in enforcement that prevents inserting anything under a node that already exists on the way down from the root to that node, this way we keep the tree minimal. This check is done by comparing `Element.less_or_equal element ancestors`: if `element` is `less_or_equal` then it doesn't bring any new information and is implied by information in `ancestors` already, the tree will not insert it.

It works well for Taint domain, however we use TreeDomain to represent a tree of tito's output paths in CollapseDepth domain and there this enforcement is causing false negatives. As it is demonstrated in `taint_in_taint_out.join_tito_collapse_test_3` test, the collapse depth tree didn't insert `[foo]: 2` branch under a root that had collapse depth `1`: because `2 >= 1` (that is `less_or_equal` in a descending integer domain) the branch was not inserted.

In fact, implication property for tito collapse depths should only hold for `0`: if there's a node with collapse depth 0, we can disregard all other children of this node (indeed, when tito is applied on this output path with 0 collapse depth, it will collapse all incoming taint and extra children on 0 node would be just repeating taint under collapsed taint and hence do not add new information). And for all other collapse depth values, we should allow children with any value, irrespecive of their `less_or_equal` relation.

To relax this property for any TreeDomain we introduce a function `transform_on_sink` that transforms elements when they are  pushed down on the way from root to the node. And a simmetrical function `transform_on_hoist`, when an element is lifted from the node to parent.

For collapse depth domain
	`transform_on_sink` should preserve 0 and map everything else to `bottom`: that would let anything but `bottom` to be inserted under any value, and nothing to be inserted under 0.
	`transform_on_hoist` keeps bottom as is and maps everything else to 0 similarly to what `approximate` does already.

Taint domain is unchanged, both functions are `id`.

Reviewed By: fahndrich

Differential Revision: D40965018

fbshipit-source-id: 9dd61c1a09ebf34555e10a7d90a9678407fd5bf6