Add tito lengths into BreadcrumbSet
Summary:
Such that any taint that is propagated (either forwardly or backwardly) via a
function that performs at least one taint transform will have a breadcrumb that
tracks the number of tito hops during the propagation. Then the breadcrumb will
be used to filter issues for easier triaging (e.g., filtering issues with a max
tito length of 3).

Implementation-wise, slot `TraceLength` of a tito taint already tracks the
number of tito hops, which is now used to build the new breadcrumb. Meanwhile,
the new breadcrumb is added only when calling a function that contains at least
one (local) taint transform.

Reviewed By: arthaud

Differential Revision: D41326072

fbshipit-source-id: 16c11e9baa7ade841e1bf2bc352de06c6dccc0ec