Inline the call to `forward_expression` in `partition_name`
Summary: In this diff, we inline a part of the code of `forward_expression` in `partition_name`. This will make explicit the call to `is_toplevel_module_reference` (and hence `resolve_exports`). This demonstrates a quadratic behavior since `resolve_exports` also visits all parents. This will allow some optimizations in further diff.

Reviewed By: grievejia

Differential Revision: D41445620

fbshipit-source-id: ddaa7999c0a7767bec5b8fc1005ad1372f2886ce