Move `partition_name` within the type checking state
Summary:
`partition_name` is a function called to partition a reference into a module name and attribute list.
It is currently implemented in the `Resolution` module, but uses `resolve_reference` which uses the type checking logic.
This diff moves `partition_name` in `typeCheck.ml`. This will allow us to call other functions from the type check logic directly.
The downside of this approach is that `partition_name` needs to be available in the `Resolution` module, so we  have to add another function "pointer" in `Resolution.t`.

Reviewed By: grievejia

Differential Revision: D41445621

fbshipit-source-id: 1afe519aafbfc100574ddc4220f2e5953b51580f