Resolve configuration path for overridden configurations
Summary:
Found a bug in the types query where the types query was not working for overridden configs.

It ended up being the case that for a pyre source `source` was being compared to `./source/...`. Our search logic relies on [this](https://fburl.com/code/3qjnxlhy) directory_contains method that failed in that specific case.

By resolving this path for the root directory, we can resolve all other paths relative to this root correctly resulting in a working query.

Reviewed By: grievejia

Differential Revision: D41448116

fbshipit-source-id: 87e8e880d570337cc6d28325bffb29462dd6f7eb