Use export resolution to speed up the partitioning of references for refinement
Summary:
`partition_name` is currently quadratic in the length of the reference, because it iterates on elements and calls `resolve_reference` which ends up calling export resolution, which also iterates on the reference.
This can be very costly for auto-generated functions with a lot of assignments (say, > 10k).

To improve the performance of this, we can run a single export resolution call, and skip iterating on the module name.

Reviewed By: grievejia

Differential Revision: D41445622

fbshipit-source-id: afc061fa942109eb22826a3fa7cdd38237e9cf3d