Support module names with many dots
Summary:
We currently support a syntax of `import a.b.c` where `a.b.c` is not necessarily an import named `c.py` within a/b or `__init__.py` within `a/b/c`, but also a file named `b.c` within the `a` directory. This diff changes the lazy module tracker to allow files with any number of dots in them.

To do that, instead of just checking in the parent directory (a.b.c -> a/b and a.b.c.extension -> a/b), we check every possible parent (a.b.c -> a/b, a, . and a.b.c.extension -> a/b, a, .). This may seem like a lot of extra work, but because it's done lazily and has a low ceiling, it should not affect perf much. It changes module resolution from being O(imports) to O(imports * max folders) which should not be very large.

Reviewed By: stroxler

Differential Revision: D41414882

fbshipit-source-id: 25361fbd18aba09f73005ef9b370b0665f7fa77f