Split the ModelParser into two modules
Summary:
This diff moves the `ModelParser.Internal` module into a different file `modelParseResult.ml`, which defines the result of parsing a `.pysa` file. This gets rid of the weird `Internal` module.

This also moves `modelQuery.ml` into `modelQueryGeneration.ml` to prevent some confusion. `ModelQuery` is actually defined in `ModelParseResult` since it is the result of parsing a file. `ModelQueryGeneration` is only responsible for executing the model query.

Reviewed By: litho17

Differential Revision: D41564613

fbshipit-source-id: e8eda8cb5f656d9216035e2fccacd6e52d05f25e