TreeDomain: `Index _` branches of the tree now shaped into `AnyIndex` of `mold`.
Summary: This diff relaxes a requirement for `mold` in `Tree.shape` to have same set of branches as in the tree, now mold can have an extra `AnyIndex` child. In this case, if `AnyIndex` is present on mold and tree has `Index _` children: they are not lifted to the parent but instead being joined (shaped) by `mold`s `AnyIndex` subtree.

Reviewed By: fahndrich

Differential Revision: D40201765

fbshipit-source-id: 3cebbcbfc0794fc8f05509d4edb7873e141c7884