launch_and_subscribe should not return a status
Summary:
In principle the `launch_and_subscribe` method of the daemon handler
could return a status used elsewhere to set the server state. But
that is not what actually happens; the `_try_connect_and_subscribe`
method handles state (which, under the current architecture, it has to because
it isn't expected to exit), and in fact the return value of
`launch_and_subscribe` is thrown away entirely. Let's just make it return
`None` instead of having a misleading type annotation and useless return
statements.

Reviewed By: grievejia

Differential Revision: D41592620

fbshipit-source-id: 9626fab66f4db4b4e5ced7ea5d299f4e76158b5f