Solve `bound variable <: Union` before solving for its `upper bound <: Union`.
Summary:
Problem: Say we solve `T (bound in a function) <: Union`. (see integration test for a snippet)

We try two approaches:
1. Try upper-bound of the LHS T <: RHS
2. Try LHS T <: each element of the union

https://www.internalfb.com/code/fbsource/[fcb9993c3beb74810d7903c6d5ef7cccb214f875]/fbcode/tools/pyre/source/analysis/constraintsSet.ml?lines=511-520

What happens is that, for a T with no explicit upper-bound (i.e., an unconstrained T), the upper-bound is `object`. So, in (1), we end up solving `object <: union`. If the Union is something like `T | None`, then this checks `object <: T`, which gives T = object.

(2) does the right thing.

However, when we have multiple solutions, we pick the first one. So, we end up returning `T = object`, which is confusing for users since the solution should be `T = T (bound)`.

Solution: As a quick fix, I'll switch the order.

I'm not sure why we do (1) at all. Will look into that later.

Reviewed By: stroxler

Differential Revision: D41671975

fbshipit-source-id: a87bebf824fe02937504339d6af30f8d0507932f