Add an integration test: Use multi-source rules to track more flows from SQL-query like strings to string operations
Summary:
To track more flows from SQL-query like strings to string operations, this diff
experiments with the approach that uses multi-source rules.

The new rule introduced in this diff is being evaluated on distillery
D41671249, to check the number of new issues and the performance degrade.

See more details from the design doc:
https://docs.google.com/document/d/1lZZtJ2ihTutLzYbIwd0ytqhDaLlE1zWh9eMduGvDoP0/edit?usp=sharing

Reviewed By: arthaud

Differential Revision: D41694209

fbshipit-source-id: 96493b92f0153eece2ae4facd2ceddc9b4e74ac6