Rename the constraint `cls.equals` into `cls.name.equals`
Summary:
For performance optimizations, we want to introduce different constraints on the name and the fully qualified name (i.e, including modules) of a function, method or class.

This is not currently possible for class constraints since the constraint is called `cls.equals`/`cls.matches`. To fix this problem, we introduce in this diff `cls.name.equals` and `cls.name.matches` with the same semantic. We will then deprecate `cls.equals` in the future.

Reviewed By: litho17

Differential Revision: D41704438

fbshipit-source-id: 64ae4343ec6036f9e5fae664a44ee683df921eb0