Rearrange code in statistics_collectory.py
Summary:
When a module defines several similar interfaces, I find it easier if the types
that are part of an interface live near the implementation rather than at
the top; it requires a bit less jumping back and forth to see how the datatypes
and logic for producing them line up.

Here I'm applying that rule to statistics_collector.py; it's just my opinion
that this is more readable so if anyone strongly disagrees I'm happy to
abandon the change.

Reviewed By: grievejia

Differential Revision: D41756715

fbshipit-source-id: 8e6ba7012673272d2c18aa8b442fc20fc5ab47f3