add support for FileOpened/FileClosed commands
Summary:
Basic idea's to start keeping a mapping of modules -> set of overlays in the server state, and keeping this mapping up to date as overlays are added and removed.

This will allow us to keep an accurate set of files for which to buck build targets in the future. In addition, we're currently running into an issue where an overlay is not created until
an edit hits due to didOpen/didClose events not getting handled even for cases where buck isn't relevant.

The spicy bit of this diff is likely the choice of data structure for maintaining open files - I chose to key it off module and represent overlays as sets. This will allow us to, say, disallow LocalUpdate's to untracked overlays in the future, but would appreciate a look from folks to make sure there's not a better place to keep track of overlays.

Reviewed By: grievejia

Differential Revision: D41615576

fbshipit-source-id: 9e94b8919e3827405ef2e6b0186dc559144b8a3b