Have separate test functions for testing attribute access.
Summary:
Problem: In `assert_attribute`, we wanted to know if the flag `accessed_through_class` is true. We did that hackily by passing in `Type[Foo]` as the `parent_instantiated_type`, checking if it was a meta-type, and then unpacking the `Type[...]`.

We can make our intent clearer by passing in a flag with a default value to make our intent clear. That also makes it easier to extend this to readonly attributes.

+ I needed to remove the `expected_attribute` keyword parameter because, otherwise, all parameters would be named and we wouldn't be able to use flags with default values.
+ Renamed `parent_instantiated_attribute` to `instantiated`.

Reviewed By: kinto0

Differential Revision: D41778056

fbshipit-source-id: fbca625fa443636c73d83aea9d92b129314a6751