Add field to indicate whether class was accessed via `ReadOnly`.
Summary:
Background: When we look up an attribute on a variable, we use `Type.class_data_from_type` to find the class.

+ `class_data_from_type` for `foo` on `int | str` would be `[ class int; class str]`
+ `class_data_from_type` for `foo` on `Type[int]` would be `[class int, accessed_through_class: true]`.

Problem: We need to recognize that looking up `foo` on `ReadOnly[int]` is a readonly type.

Solution: This is similar to `Type[int]`. `class_data_from_type` for `foo` on `ReadOnly[int]` should be `[ class int, attribute foo, accessed_through_class: false, accessed_through_readonly: true ]`.

This diff just adds the record field and always sets it to `false`. Upcoming diffs will change it and update callers.

Reviewed By: sinancepel

Differential Revision: D41778052

fbshipit-source-id: 3a6685e2b2f480a45031cca858d1b20e4627c452